Написал контроллеры на методы в интерфейсах не написаны методы в AdsController (deleteAd и updateAdImage) так как не получилось в них разобраться

+++ src/main/java/ru/skypro/homework/service/AvatarService.java
@@ -5,5 +5,5 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 
 public interface AvatarService {
-    void updateAvatar(Integer id, MultipartFile image) throws IOException;
+    void updateAvatar(Long id, MultipartFile image) throws IOException;
 }
